+++ ironforge-component/ironforge-base/base-server/src/main/java/org/ironforge/base/persist/repo/CountryRepository.java
@@ -5,11 +5,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CountryRepository extends JpaRepository<TCountry3166, Long> {
+public interface CountryRepository extends JpaRepository<TCountry3166, Integer> {
 
-  TCountry3166 findByAlphaCode2(String alphaCode2);
+    TCountry3166 findByAlphaCode2(String alphaCode2);
 
-  TCountry3166 findByAlphaCode3(String alphaCode3);
+    TCountry3166 findByAlphaCode3(String alphaCode3);
 
-  TCountry3166 findByNumberCode(String numberCode);
+    TCountry3166 findByNumberCode(String numberCode);
 }
