+++ ironforge-component/ironforge-base/base-server/src/main/java/org/ironforge/base/persist/repo/MccRepository.java
@@ -3,7 +3,7 @@ package org.ironforge.base.persist.repo;
 import org.ironforge.base.persist.entity.TMccDef;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MccRepository extends JpaRepository<TMccDef, Long> {
+public interface MccRepository extends JpaRepository<TMccDef, Integer> {
 
   TMccDef findByMccCode(String mccCode);
 }
